Drop check of package in sshd_required definitions
This is not the best place to check if openssh-server is installed.

We can check for openssh-server package when sshd is required and not
required.
But when sshd_required is not set, we don't check if openssh-server is
installed or not, because both are valid states.

This gives the impression that when extending sshd_required_or_unset
and sshd_not_required_or_unset there is no need to check for
openssh-server package, which is not true.

The only purpose of these definitions should be to check for state of
sshd_required value.