Improve check and remediation for GNOME banner (RHEL 7)
This commit adds a bash remediation script for the rule
dconf_gnome_login_banner_text. Since, the banner format is different from the
one used by /etc/issue, it has been necessary to add a specfic regular
expression in order to handle that case. Indeed, the banner format in dconf must
be with \n and not with real new lines.

That's why a new value has been added : dconf_banner_text with a regular
expression matching that requirement. The current check has been updated in
order to use that variable in stead of login_banner_text.

Check and new bash remediation use this new value in stead of the
login_banner_text one used for /etc/issue.