Stop shadowing package_prelink_removed
The shared/checks/oval/package_prelink_removed.xml file was shadowed
in RHEL7 and SLE12 by OVAL generated from template. In Fedora and
RHEL6, the shared version was used. We remove the OVAL from shared
and generate it in Fedora and RHEL6 from template as well. We remove
it from SLE12, because there is no rule in SLE12 checking for prelink
package.