Add uid_min to product.yml and derive auid from it
uid_min defines the starting value for user ids in the platform.
If not set in product.yml, its value will 1000.

The auid value is used in audit rules to set which user processes to collect data from.
The auid value in the rules should be set to uid_min.