Do not ignore failures of external commands
The CalledProcessError exception was caught and the return value of
this function was ignored. That meant if some command went wrong
(eg. uploading the tarball to the VM was not successful or extraction
of the tarball was not successful) it was ignored and the test suite
continued happily, without letting the user know. We need to terminate
if some of the test harness setup tasks went wrong.