Remove invalid link
This link goes to fedorahosted retired page. It's moved to GitHub,
but I'm not sure if I have found the right place in code:
https://github.com/linux-audit/audit-userspace/blob/ef1152ada432ca85b19e92122c40af6df7bf10d9/common/strsplit.c#L71
I would prefer to remove the link because the place in code can easily
change over time and we don't have capacity to watch the changes in
auditd sources.