Add missing quotes
Addressing:
```
ERROR! Syntax Error while loading YAML.
  found unacceptable key (unhashable type: 'AnsibleMapping')

The error appears to have been in
'/home/jenkins/workspace/scap-security-guide-pull-requests/build/roles/ssg-rhel7-role-nist-800-171-cui.yml':
line 1174, column 17, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

        option: memcache_timeout
        value: {{ var_sssd_memcache_timeout }}
                ^ here
We could be wrong, but this one looks like it might be an issue with
missing quotes.  Always quote template expression brackets when they
start a value. For instance:

    with_items:
      - {{ foo }}

Should be written as:

    with_items:
      - "{{ foo }}"
```