Enable OVAL for rule grub2_enable_fips_mode on Fedora
I change the condition that the installed OS must be RHEL 7
to a check that installed OS is certified, because there
can be multiple operating systems that are FIPS certified,
not only RHEL 7. However, the rule will always fail on
Fedora.  I also add Fedora to the disclaimer.  I had to
add `multi_platform_fedora` to checks and remediations for
rule `package_dracut-fips_installed` which is a dependency,
otherwise the OVAL definition `package_dracut-fips_installed`
that extends this OVAL definition would be missing and that
would lead to the fact that the whole OVAL definition will
be stripped by our build system.