Add profiles to man page automatically
The list of profiles in man page gets outdated, because people sometimes
don't add new profiles or products to man page. Also they don't update
their descriptions in the man page. We would like to avoid outdated
information and we would like to avoid manual work. I propose to
generate the list of profiles and their descriptions by an automated
script which will be run as a part of build process. This will also
solve the problem that we have on RHEL where the man page describes
Fedora profiles which are not shipped. Only the products that are
built will be added to man page.