Added support for @ANSIBLE_ENSURE_PLATFORM@.
Some rules can define platforms that they are applicable for
using the `<platform idref="cpe:/a:platform-id"/>` tag.
The @ANSIBLE_ENSURE_PLATFORM@ macro expands to a when: ... Ansible clause,
if the respective rule is bound to a platform.

- If a known platform is encountered, respective `when` entry is created.
- If an unknown platform is encountered,
  a descriptive error comment is created.
- If no platform is specified, an informative comment is created.

Only one platform per rule is supported at the moment.
The 'machine' platform is supported.