Added support for default role variables.
The script now generates `when:` clauses to each task using newly introduced variables.
Those variables are extracted from existing tags. For every variable added, a corresponding entry
in `defaults/main.yml` is created, where the variable is set to true,
so the playbook behaves in the same way as before this change.