Check if external variable has the expected value
This commit adds a new criterion to OVAL definition for rule
`enable_fips_mode`. The test checks if `var_system_crypto_policy` is set
to FIPS.  This solves the situation when user tailors his profile and
refines the XCCDF value `var_system_crypto_policy` to a different value
than FIPS. In this situation the rule passes, although the FIPS mode is
not enabled. It is obvious that this tailoring does not make any sense,
but we should be more anticipatory.