Remove unused code
The env_yaml variable in main() is set but not used anywhere.  The
env_yaml can be removed, which means the --build_config_yaml and
--product-yaml arguments will not be used, so they don't have to be
provided.  When we remove --build-config-yaml and --product-yaml we need
to change also the CMake command that calls this script.  Also, re
module is not used anywhere in this script, so we don't have to import
it.