Simplify build_rule_playbooks.py CLI
It isn't recommended to require `--input-dir` and `--output-dir`,
because in their typical usage in ComplianceAsCode build system
in CMake, their value can be computed from the CMAKE_SOURCE_DIR
and the ID of the product we want to build Playbooks for.
We will keep `--input-dir` and `--output-dir` for users who want
to use the script stand-alone, eg. for development purposes.
Fixes #4034.