Build playbooks also for virtual '(all)' profile
Generates playbooks for all product rules without profile context to a
directory.
A product typically contains lots of rules, not all of them are part of
any profile. As it may happen that a rule falls out of a profile, it is
a good idea to be able to generate playbooks for all rules that are not
part of any profile using the build_rule_playbooks.py script.