Change directory names
Playbooks go to `/build/ansible`, bash scripts go to `/build/bash`.
This will have a benefit that it's consistent with dowstream packaging.
For example in Fedora we have the playbooks in
`/usr/share/scap-security-guide/ansible/` so it would be nice to have them
in `/build/ansible` in upstream as well.