install_vm.py: redesign --distro as URL alias
Libvirt can autodetect --variant on --location=url installs from
the downloaded kernel/initrd.

This works even on non-RPM distributions, even though the matching
isn't as reliable as with Fedora-based distros.

However the worst case is a default "good enough" fallback, which
should be more than enough for our use case here.

The --distro is then left only as an --url alias, for convenience.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>