Workaround SCAPVal 1.3.2 NullPointerException
If run as a non-root user SCAPVal 1.3.2 can't generate HTML report
because it throws a NullPointerException. This bug has been reproduced
on Fedora 29 and RHEL 7.  We don't wish to run this test as root user.
But, we don't need the HTML report for this test, because we parse the
plain XML results. For the purpose of this test we can ignore this
SCAPVal error until SCAPVal is fixed.