Check if all rules in profile exist
Prints an error message and aborts the build if profile selects
a rule that doesn't exist or isn't present in the built product.
The check is done early in the build process which means users
won't have to validate the built content to find mistakes such
as typo in rule ID in the profile file.
Fixes: #4541