Use block style when a string contains a newline
By default, the yaml.Dumper dumps multiline strings surrounded by single
quotes. If a newline character is a part of the string it is represented
by an empty line in a single quoted string. The line breaks in a single
quoted string are replaced by space. This assumes the user needs
to know YAML specification to understand the behavior of the single
quoted string. Also it looks ugly. Instead, we can use a literal
block style, represented by '|' character. In literal blocks, the line
breaks are preserved and no extra empty lines are needed.