Simplify loading of Jinja macros
There's no real reason to return a new dictionary instead of extending
the existing dictionary when we're going to end up modifying the
existing dictionary at some point anyways.

At best this helped to not pollute substitutions_dict when an exception
occurred, however exceptions from this function are usually passed back
up to CMake and not otherwise handled.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>