Test suite auto-detect datastream.
The frequent use case is that a single datastream is built in the build directory,
and one wants it to be tested by the test suite.
This change add autodetection capabilities, so the --datastream option can be left out.
If it is left out and there is not a single datastream in the build dir,
an error is displayed instead.