Convert functions from include_lineinfile.sh to Jinja macros
Currently, the Bash remediation functions are saved in an XCCDF Value
and their code is substituted by OpenSCAP when the remediation is
executed.  However, the way of using the functions in individual
remediation is tedious and hard to understand. People have to source a
non-existent file /usr/share/scap-security-guide/remediation_functions
and then call an empty function to make the function code available in
the remediation. We have discovered that it does not allow to use
multiple functions in a single remediation. Another drawback is that
the function code is duplicated in HTML report each time it is used
which makes the report bigger. This patch aims to simplify creating the
Bash remediation and address the aforementioned problems by converting
the functions to Jinja macro. In Bash remediation you can now simply
call a Jinja macro and a simple, stand-alone Bash code will be generated
for you.  In this patch only the functions from include_lineinfile.sh
are converted to Jinja. Other shared modules can be converted in future
in a similar way.