Process template arguments for each template separately
We sometimes need to process arguments for OVAL differently than for
remediations. For example, in current templating system, in
`create_audit_rules_usergroup_modification.py`, we escape backlashes in
`path` argument for OVAL template, but not for other templates. There
are couple more examples. Therefore we need to process the arguments
separately for each template. The callback functions can be
parametrized based on the target language.