Change the template arguments keys
If we will change the keys to uppercase at the latest moment, we can use
the lowercase keys in callback, which means we will not have to change
the callbacks code when the variables in templates will be converted to
lowercase.

Keep in mind that the changing the keys to uppercase is a temporary
measure and it will be removed at the moment the variables in Jinja
templates are converted to lowercase.