Make template_OVAL_service_disabled self-contained
The template template_OVAL_service_disabled used extend_definition
which referenced OVAL definition generated from different template
template_OVAL_package_removed. With the new templating system, we
generate templates based on rules. But for certain packages s.a.
acpid we don't have rules that check if the package is removed.
In old system it was enough to just declare a line in CSV, but in
new system we would have to create the rule and make sure it is
present in all products as the dependent rule. It would be cumbersome
to explain this to content contributors.

Instead, we provide a more user friendly solution. The
template_OVAL_service_disabled will not depend on any other generated
content, it will contain all the required checks in a single file.
That means incorporating tests of presence of a RPM/DEB package into
the template. The template won't use extend_definition, but there is
a new criterion instead.

The drawback of this patch is that the result OVAL content blows up
in size.

To prevent copying the code, the common code for both templates has
been refactored out to a Jinja macro.