Enable permissions checks of directories
Adds logic that will ensure if the value of `filepath` key ends with a
`/` the OVAL check for directories would be generated. This is because
the `filepath` element in OVAL file_object can't be used to describe
directories (see OVAL Unix file_object specification). To describe
directory you have to use `path` element instead and set `xsi:nil`
attribute of filename element to `true`.