Fix templates file_owner and file_groupowner
During the port of template data to CSVs we have forgotten about 4 rules
in OCP3 benchmark which check for owner and group owner of files
specified by a wild card. The result was that we haven't port the
feature that enabled generating this type of check as we incorrectly
thought it isn't used. To avoid regression in OCP3 Benchmark after the
new templating system is switched on, we need to reintroduce this
feature in templates `file_owner` and `file_groupowner`. During the
migration of the data the base name and directory path has been merged.
So we propose an alternative way: there will be a new optional variable
`file_regex` in the template which can contain a regular expression
describing files in a given directory.  If `file_regex` is set the
`filepath` must describe a directory, which we recognize by the trailing
backslash.