Fix regex for OpenShift auditConfig related checks
The checks for the audit log related settings of the OpenShift API
server have an extra ':' character where it attempts to match the
'auditConfig' string.  This patch fixes the 3 affected checks to
expect only a single ':' character.  In addition, a 4th audit log
check was updated to remove unneccessary escaping to be consistent.