SSGTS: Use wildcards instead of matching substring
This commit changes the way test suite users specify rules to be tested
in rule mode. Instead of matching a substring in rule ID, we will accept
wildcards as rule mode arguments. The wildcards are similar to shell
globs, and use Python fnmatch library. For example, argument
accounts_passwords_pam_faillock_deny will run test scenarios for a
single rule and accounts_passwords* will run test scenarios for all
rules which ID starts with accounts_passwords.

The rule ID prefix (xccdf_org.ssgproject.content_rule_) does not need to
be provided by the user because it will be added automatically.

Using substring match to find rules to be tested caused problems in
situation when we have 2 rule IDs and one of them is a substring of
another. For example, when an user requested to run test scenarios for
rule accounts_passwords_pam_faillock_deny, the test suite run tests also
for rule accounts_passwords_pam_faillock_deny_root. That was annoying
because testing the latter takes some time.