Check EKU in rsyslog remote configuration
This patch adds option "streamdriver.CheckExtendedKeyPurpose" to the
list of mandatory settings in rule rsyslog_remote_tls and requires this
option to be set to "on".  This option configures rsyslog to check for
the ExtendedKeyUsage field of the certificate.  This is a new feature in
rsyslog introduced in https://github.com/rsyslog/rsyslog/pull/3845

Fixes: #5098