Create script to build datastream images from source
This script runs the build_product script (taking the product as a
parameter), and it pushes the ssg-$PRODUCT-ds.xml file to a container.

The container is created by using a BuildConfig and an ImageStream,
these are updated by pushing the local xml file to the container build.

This is just one step of the build process which helps with local
development. In further iterations I'll create a BuildConfig that builds
inside a container as opposed to requiring a local file.

Example:

    ./utils/build_ds_container.sh ocp4