utils: A helper script to generate remediations for files
It is tedious to craft the ignition remediations for files, mostly due
to having to urlencode the file input. What is even more difficult is to
see the resulting file.

This script attempts to help with both by offering an encode and decode
actions that can either generate an ignition file based on a prepared
file or printing out how the file would look like based on the ignition
snippet.