Make banner compatible with console and dconf
The banner in /etc/issue is expected to have actual newlines, while the
banner in /etc/dconf/db/gdm.d/ is expected to have the escape sequence
'\n'.

This commit transforms the newline from the input banner into a regex
that matches either the newline or the escape sequence.

During remediation, each rule will replace the regular expression for
the correct "version" of the newline.