ocp4: reset client in e2e tests after installing operator
This client reset is needed because the available resources in the
cluster are cached y a caching client, but when this client starts
doing operators, the operator hasn't been installed.

What results from this is that the compliance-operator's resources can't
be used by the time we try to test.

To address this, we reset the cache after installing the operator.