ocp4: fix image pullspec output from build_ds_container.sh
It was assuming there would only be one imagestream in the namespace.
But this is not necessarily the case. So this calls specifically for the
imagestream that's being built and only outputs the pullspec when the
build is complete.

For slow connections, this also waits now 5 seconds for the build to
appear in k8s

Finally, this adds the namespace parameters to all `oc` calls, so this
can be called from another namespace and still work.