shared/partition.sh: Increase the size of a test device
In some test scenarios (`fstab.fail.sh`) we create a new device
with size of 50MB to test rule `mount_option_tmp_noexec` and it
is re-mounted as the `/tmp`. This causes that there is not enough
space in `/tmp` where our test suite produces logs and result files.
This commit increases the size of the created device so logs can
fit in. Logs are copied from the guest to the host before the testing
device mounted as `/tmp` is unmounted and removed so logs are not lost.