no_shelllogin_for_systemaccounts: simplify check for range of UIDs
There is no need to make calculations on top of the UIDs, we can compare
the collected UIDs with shell againt the states that define the valid range.

Avoiding the calculations has the added benefit of not using/referencing
a variable that can be empty (when no user has shell, except root).