Use specific OVAL check to verify that a system is RHCOS4
We used to bundle the check into the RHEL8 checks. However, we've
noticed that the same checks for RHCOS won't be the same as for RHEL. So
let's separate them.

On the other hand, the RHCOS checks were broken by a recent change on
the `PRETTY_NAME` format in the `/etc/os-release` file we were checking.
So instead, we switch to checking the `ID` key to match `rhcos` and
`VERSION_ID` to match `4`.