Change oval check to verify if we're in OCP4
The check used to verify if it was a RHEL node, but this is not ideal.
This changes it so that we rely on an object fetched from the cluster
that contains the version, and we check that the version is correct.

This effectively makes the OCP4 product only usable for API/platform
checks (meaning, kubernetes checks).