Introduce platform_package_overrides
Introduce a mapping of CPE package platform name to a package name.

Each linux distro or version may have its specific name for a package,
this mapping allows a product to override the package name of a
platorm.

By default, it assumes that the package name will be the same as the
platform name.