ocp4: Add option to create profile bundles for the build_ds_container.sh script
ProfileBundles are how the compliance-operator is able to track content
images, parse them and expose the profiles and rules to users. This adds
a small piece of automation, when building content images for OCP, to be
able to create the relevant profilebundles when running that script.
This would be useful when developing content.