Add a test for missing CCEs
This test checks if all rules that are a part of a profile have a CCE
identifier properly assigned. It checks the built datastream. It will
check RHEL 6, RHEL 7 and RHEL 8. The test is a part of CTest. It will be
executed on every PR. It will prevent introducing rules without CCEs.