Add placeholders for CIS profile checks in ocp4
These are just comments which will eventually have the checks besides
them. This is done in preparation for this work.

The existing rules have been placed in order, under the comment where
the rule should be. Note that they currently don't work on OCP4 and will
be fixed as part of the hackathon.