ocp4/e2e - WORKAROUND: Use suffix to detect scan type
The e2e tests were written before we had scansettingbinding objects
created. Hence, it uses the benchmark directory to determine the scan
type.

This takes into account the "-node" suffix to determine if it's a node
or a platform scan. This way, we can make the cis-node profile work.

In the long run, we should rewrite the tests to stop using
ComplianceSuites directly and use ScanSettingBindings instead.