OCP4: Enhance e2e tests to check individual rules
This enhances the e2e tests to be able to checkf or individual rules.

This requires folks to create a `/tests/ocp4` directory in the rule
that's to be tested. If the directory doesn't exist, the verification
will be skipped.

If the directory exists, it'll then look for a `e2e.yml` file that
contains the test definition.

The current format is:

```
---
default_result: <result>
result_after_remediation: <result>
```

With result being either `PASS` or `FAIL`.

If there were remediations applied the test will either check the value
of `result_after_remediation` if it's not empty or `default_result`.

This enables us to test that rules don't regress and that they issue the
expected results.