OCP4/CIS 2.7: Check that etcd has its own CA
This introduces a rule that checks that etcd is using its own CA which
is different from Kube's CA. It does this by using custom OVAL to check
that the two CA files are indeed different.

Since both components (etcd and the kube apiserver) only run on the
master nodes, this introduces a new cpe `ocp4-master-node`. This allows
us to differentiate between master nodes and other nodes. However, take
into account that it's on a node level and it isn't applicable to
"platform checks"