Stop calculating the time difference to current time
Calculating the time difference of both files against current time and
comparing them is imprecise, as both calculations of the time difference
can not be done simultaneously.
One of the files will always appear to be a bit older than the other.

If both files were generated one right after the other (i.e. on the
same second), but calculation of difference against current time was
done in different units of "second". The test will result in a false
negative.

Lets just compare their age in seconds since epoch directly.