Add more parameters to sysctl template
To check sysctl value net.ipv4.ip_local_port_range we need to
modify the template. This value consists of two integers. The OpenSCAP
probe and also sysctl command returns these 2 integers separated by a
tab. OVAL doesn't have a tuple data type so we represent the value as a
string. The string contains the 2 values separated by a space. But
to match the sysctl item, we should rather use a regular expression.
We still need the space-separated version for the remediation, though.
Therefore we can extend the template to allow specify pattern match
operation and to allow a regular expression value.